+++ CHANGELOG.md
@@ -5,17 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.1.3] - 2026-02-
+## [v1.1.4] - 2026-03-19
 
 ### Added
 
--
+- MkDocs configuration for documentation site deployment to GitHub Pages
+
+### Changed
+
+- Updated documentation and docstrings to reflect recent changes and improvements
+
+## [v1.1.3] - 2026-03-16
+
+### Added
+
+- Support for provider classes in LazyWiring, allowing for more flexible and intuitive dependency injection configurations
+- Fallback plugin will be used as parent for all orphans injectables, allowing for better handling of unregistered dependencies
 
 ### Changed
 
 - LazyWiring now accepts provider classes directly, allowing for more intuitive usage
 
-## [v1.1.2] - 2026-02-
+## [v1.1.2] - 2026-03-16
 
 ### Added